backend/requirements-vercel-light.txt
fastapi==0.115.6
uvicorn[standard]==0.24.0
pydantic>=2.10.3
python-dotenv==1.0.0
pydantic-settings==2.1.0
requests==2.31.0
python-multipart==0.0.6
mangum==0.17.0